Fix: 회원 정보 수정 로직 수정
#288

+++ src/main/java/com/zerobase/foodlier/global/profile/facade/ProfileFacade.java
@@ -30,7 +30,8 @@ public class ProfileFacade {
         String imageUrl = form.getProfileImage() != null ?
                 s3Service.getImageUrl(form.getProfileImage()) :
                 member.getProfileUrl();
-        if (!imageUrl.equals(member.getProfileUrl())) {
+        if (!imageUrl.equals(member.getProfileUrl())
+                && member.getProfileUrl() != null) {
             s3Service.deleteImage(member.getProfileUrl());
         }
         memberService.updatePrivateProfile(
